#171 : move Slot to other class, and added a tests for it, also did a little refactoring
